+++ geo/src/algorithm/cross_track_distance.rs
@@ -4,6 +4,8 @@ use num_traits::FromPrimitive;
 
 /// Determine the cross track distance (also known as the cross track error) which is the shortest
 /// distance between a point and a continuous line.
+///
+/// Distance is measured in meters, using the Haversine formula.
 pub trait CrossTrackDistance<T, Rhs = Self>
 where
     T: CoordFloat + FromPrimitive,
@@ -37,31 +39,22 @@ where
     ///     distance.round()
     /// );
     /// ```
-    fn cross_track_distance<MetricSpace>(&self, line_point_a: &Rhs, line_point_b: &Rhs) -> T
-    where
-        MetricSpace: for<'a> Distance<T, &'a Line<T>, &'a Point<T>>;
+    fn cross_track_distance(&self, line_point_a: &Rhs, line_point_b: &Rhs) -> T;
 }
 
 impl<T> CrossTrackDistance<T, Point<T>> for Point<T>
 where
     T: CoordFloat + FromPrimitive,
 {
-    fn cross_track_distance<MetricSpace>(
-        &self,
-        line_point_a: &Point<T>,
-        line_point_b: &Point<T>,
-    ) -> T
-    where
-        MetricSpace: for<'a> Distance<T, &'a Line<T>, &'a Point<T>>,
-    {
-        MetricSpace::distance(&Line::new(*line_point_a, *line_point_b), self)
+    fn cross_track_distance(&self, line_point_a: &Point<T>, line_point_b: &Point<T>) -> T {
+        crate::Haversine::distance(&Line::new(*line_point_a, *line_point_b), self)
     }
 }
 
 #[cfg(test)]
 mod test {
     use crate::Point;
-    use crate::{CrossTrackDistance, Distance, Euclidean, Haversine};
+    use crate::{CrossTrackDistance, Distance, Haversine};
 
     #[test]
     fn distance1_test() {
@@ -69,7 +62,7 @@ mod test {
         let line_point_a = Point::new(-1.7297, 53.3206);
         let line_point_b = Point::new(0.1334, 53.1887);
         assert_relative_eq!(
-            p.cross_track_distance::<Haversine>(&line_point_a, &line_point_b),
+            p.cross_track_distance(&line_point_a, &line_point_b),
             307.549995,
             epsilon = 1.0e-6
         );
@@ -82,7 +75,7 @@ mod test {
         let line_point_b = Point::new(2., 0.);
 
         assert_relative_eq!(
-            p.cross_track_distance::<Haversine>(&line_point_a, &line_point_b),
+            p.cross_track_distance(&line_point_a, &line_point_b),
             0.,
             epsilon = 1.0e-6
         );
@@ -95,13 +88,13 @@ mod test {
         let line_point_b = Point::new(1., 1.);
 
         assert_relative_eq!(
-            p.cross_track_distance::<Haversine>(&line_point_a, &line_point_b),
+            p.cross_track_distance(&line_point_a, &line_point_b),
             Haversine::distance(p, Point::new(1., 0.)),
             epsilon = 1.0e-6
         );
 
         assert_relative_eq!(
-            p.cross_track_distance::<Haversine>(&line_point_b, &line_point_a),
+            p.cross_track_distance(&line_point_b, &line_point_a),
             Haversine::distance(p, Point::new(1., 0.)),
             epsilon = 1.0e-6
         );
@@ -113,15 +106,7 @@ mod test {
         let line_point_a = Point::new(-80.1918f64, 25.7617f64);
         let line_point_b = Point::new(-120.7401f64, 47.7511f64);
 
-        let haversine_distance = p1.cross_track_distance::<Haversine>(&line_point_a, &line_point_b);
+        let haversine_distance = p1.cross_track_distance(&line_point_a, &line_point_b);
         assert_relative_eq!(haversine_distance, 1_547_104., epsilon = 1.0);
-
-        // Same as above, but projected to EPSG:5070 to test Euclidean
-        let p1 = Point::new(1826303.9422258963, 2179112.0732980534);
-        let line_point_a = Point::new(1594077.349564382, 434470.0414052719);
-        let line_point_b = Point::new(-1847681.2454118263, 2992574.0850278544);
-
-        let euclidean_distance = p1.cross_track_distance::<Euclidean>(&line_point_a, &line_point_b);
-        assert_relative_eq!(euclidean_distance, 1_538_764., epsilon = 1.0);
     }
 }